+++ core/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `chg` **Separate Pakyow's logging concerns from that of async.**
+
+    *Related links:*
+    - [Pull Request #540][pr-540]
+
   * `fix` **Maintain memory database connections when forking.**
 
     *Related links:*
@@ -755,6 +760,7 @@
     *Related links:*
     - [Pull Request #338][pr-338]
 
+[pr-540]: https://github.com/pakyow/pakyow/pull/540
 [pr-536]: https://github.com/pakyow/pakyow/pull/536
 [pr-535]: https://github.com/pakyow/pakyow/pull/535
 [pr-533]: https://github.com/pakyow/pakyow/pull/533